refactor: dividing code
